chromium/third_party/blink/web_tests/external/wpt/webrtc/RTCRtpTransceiver.https-expected.txt

This is a testharness.js-based test.
PASS checkAddTransceiverNoTrack
PASS checkAddTransceiverWithTrack
PASS checkAddTransceiverWithAddTrack
PASS checkAddTransceiverWithDirection
FAIL checkMsidNoTrackId promise_test: Unhandled rejection with value: object "OperationError: Failed to execute 'setRemoteDescription' on 'RTCPeerConnection': Failed to parse SessionDescription. a=msid:fake-stream-id Expects 2 fields."
PASS checkAddTransceiverWithSetRemoteOfferSending
PASS checkAddTransceiverWithSetRemoteOfferNoSend
PASS checkAddTransceiverBadKind
PASS checkNoMidOffer
PASS checkNoMidAnswer
PASS checkSetDirection
PASS checkCurrentDirection
PASS checkSendrecvWithNoSendTrack
PASS checkSendrecvWithTracklessStream
PASS checkAddTransceiverNoTrackDoesntPair
PASS checkAddTransceiverWithTrackDoesntPair
PASS checkAddTransceiverThenReplaceTrackDoesntPair
FAIL checkAddTransceiverThenAddTrackPairs assert_equals: expected "[{sender:{track:{}}}]" but got "[{sender:{track:{}}},{}]"
PASS checkAddTrackPairs
PASS checkReplaceTrackNullDoesntPreventPairing
PASS checkRemoveAndReadd
PASS checkAddTrackExistingTransceiverThenRemove
FAIL checkRemoveTrackNegotiation assert_equals: pc2.setRemoteDescription(offer) should've added 2 tracks to receive stream expected 2 but got 0
FAIL checkMute assert_true: expected true got false
FAIL checkStop promise_test: Unhandled rejection with value: object "TypeError: stoppedTransceiver.stop is not a function"
FAIL checkStopAfterCreateOffer promise_test: Unhandled rejection with value: object "TypeError: pc1.getTransceivers(...)[0].stop is not a function"
FAIL checkStopAfterSetLocalOffer promise_test: Unhandled rejection with value: object "TypeError: pc1.getTransceivers(...)[0].stop is not a function"
FAIL checkStopAfterSetRemoteOffer promise_test: Unhandled rejection with value: object "TypeError: pc2.getTransceivers(...)[0].stop is not a function"
FAIL checkStopAfterCreateAnswer promise_test: Unhandled rejection with value: object "TypeError: pc2.getTransceivers(...)[0].stop is not a function"
FAIL checkStopAfterSetLocalAnswer promise_test: Unhandled rejection with value: object "TypeError: pc2.getTransceivers(...)[0].stop is not a function"
FAIL checkStopAfterClose assert_equals: Stopping a transceiver on a closed PC should throw. throws InvalidStateError expected "InvalidStateError" but got "TypeError"
FAIL checkLocalRollback promise_test: Unhandled rejection with value: object "TypeError: pc.getTransceivers(...)[0].stop is not a function"
PASS checkRollbackAndSetRemoteOfferWithDifferentType
FAIL checkRemoteRollback assert_equals: expected "{currentDirection:null,mid:null,stopped:true}" but got "{currentDirection:null,mid:null,stopped:false}"
FAIL checkMsectionReuse promise_test: Unhandled rejection with value: object "TypeError: pc2.getTransceivers(...)[0].stop is not a function"
FAIL checkStopAfterCreateOfferWithReusedMsection promise_test: Unhandled rejection with value: object "TypeError: pc1.getTransceivers(...)[1].stop is not a function"
FAIL checkAddIceCandidateToStoppedTransceiver promise_test: Unhandled rejection with value: object "TypeError: pc1.getTransceivers(...)[1].stop is not a function"
FAIL checkBundleTagRejected promise_test: Unhandled rejection with value: object "TypeError: pc2.getTransceivers(...)[0].stop is not a function"
Harness: the test ran to completion.
